+++ src/main/java/com/xiaoleilu/hulu/servlet/DefaultServlet.java
@@ -1,7 +1,8 @@
-package com.xiaoleilu.hulu;
+package com.xiaoleilu.hulu.servlet;
 
 import java.io.IOException;
 
+import javax.servlet.Servlet;
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -9,32 +10,36 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * Servlet 入口<br>
- * 此入口如果拦截所有请求，静态文件将无法找到
+ * 默认Servlet<br>
  * @author Looly
  *
  */
-//@WebServlet(urlPatterns="/*", loadOnStartup=1)
-public class ActionServlet extends HttpServlet{
-	private static final long serialVersionUID = -1429875457807724161L;
+public class DefaultServlet extends HttpServlet{
+	private static final long serialVersionUID = -8137983239382306470L;
+	
+	private Servlet targetDefaultServlet;
 
 	@Override
 	public void init(ServletConfig config) throws ServletException {
-		ActionContext.init(config.getServletContext());
+		try {
+			targetDefaultServlet = new org.apache.catalina.servlets.DefaultServlet();
+		} catch (Exception e) {
+			targetDefaultServlet = new org.eclipse.jetty.servlet.DefaultServlet();
+		}
+		targetDefaultServlet.init(config);
 		super.init(config);
 	}
 	
 	@Override
 	protected void service(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
-		//-- 处理请求，如果处理失败（无对应的Action），执行父类逻辑
-		if(false == ActionContext.handle(req, res)) {
+		if(null != targetDefaultServlet){
+			targetDefaultServlet.service(req, res);
+		}else{
 			super.service(req, res);
 		}
 	}
 	
 	@Override
 	public void destroy() {
-		super.destroy();
-		ActionContext.destroy();
 	}
 }